refactor(config): split configuration responsibilities
